Update
Se separan clases y se optimiza un poquito 💯
