+++ app/src/main/java/com/example/karen_and/models/Content.kt
@@ -1,0 +1,7 @@
+package com.example.karen_and.models
+
+data class Content(
+    val id: Int,
+    val classId: Int,
+    val content: String
+)
