+++ app/src/main/java/com/example/karen_and/models/LessonStatus.kt
@@ -1,0 +1,6 @@
+package com.example.karen_and.models
+
+data class LessonStatus(
+    val id: Int,
+    val name: String
+)
