+++ app/src/main/java/com/example/karen_and/models/Subject.kt
@@ -1,0 +1,8 @@
+package com.example.karen_and.models
+
+data class Subject(
+    val id: Int,
+    val name: String,
+    val userId: Int,
+    val lessons: List<Lesson>
+)
